+++ app/src/main/java/woowacourse/shopping/data/local/entity/PurchaseProductEntity.kt
@@ -1,4 +1,4 @@
-package woowacourse.shopping.data.entity
+package woowacourse.shopping.data.local.entity
 
 import androidx.room.ColumnInfo
 import androidx.room.Entity
